extract memu configuration to own module
